Do not errback the polling Deferred in connectionLost.
Doing so prevents completing the cycle where we call call poll()
repeatedly until we get a success or an error result from psycopg2. If
this cycle is broken by prematurely errbacking the polling Deferred,
psycopg2 is left in a state where it things there still is an
asynchronous query being executed. Based on the bug report and test
case by Peter Le Bek.

Fixes #1.