Do not try to close the cursor after errors.
If there is a problem with the connection, trying to close the cursor
will just result in another error which might mask the original one,
as demonstrated by the new regression test that simulates a backend
failure. In any case, closing the cursor is not really necessary and
has almost no effect on unnamed cursors in psycopg2.