Fix racy cancellation test.
The test was cancelling an Interaction, which was doing some
operations first and then executing a very long query., which was
supposed to be cancelled. It turns out that waiting an arbitrary
amount of time before issuing the cancellation was too racy, as the
initial operations might have taken so long that it was one of them
that was being cancelled, leaving the long-running one and hanging the
test.

Fix it by rewriting the test to only schedule cancellation from inside
of the interaction only after the initial preparation is done.

Bug report and initial investigation by Lex Berezhny.