Fix a bug where a connection attempt would be left hanging.
Happens if the server is not listening. The first status returned by
psycopg2 is POLL_WRITE, so txpostgres waits for the socket to become
writable. The moment the socket gets added to the file descriptor set
being watched, it triggers an error (since the connection got refused)
and connectionLost gets called. This left the Deferred returned from
the first poll() hanging. What's interesting is that Twisted versions
up to and including 11.0.0 only exhibited the buggy behaviour with the
epoll reactor. Twisted newer than 11.0.0 makes the bug manifest with
every reactor.

Fix this by continuing to poll even after the connection gets lost and
rely on psycopg2 to detect the condition and raise an exception when
polled. Be careful not to continue polling if the connection had been
closed by the user, though.

Per bug report and patch from Vasily Evseenko and Andrey Smirnov. The
committed version is a bit different from what they proposed, but
based on the same basic idea.

Fixes #4.