Do not set the pollables to None immediately before closing them.
Instead, rely on the closed property to determine whether it is safe
to call poll() on the pollable object. That's more explicit than
relying on the pollables being None and more correct, too, because
removing the references to the pollable immediately after closing it
would lead to errors such as reported in issue #5.

It is not certain that this commit fixes the issue, but it definitely
is a step forward from the code quality standpoint.