Explicitly call connectionLost when closing the connection.
When using the epoll reactor, file descriptors that get closed are
automatically removed from the epoll set. This can cause Twisted to
never call connectionLost of objects that use this file
description. Missing connectionLost can cause Deferreds returned to
callers to never be errbacked, leaving the callers hanging.

Fix this by tracking the cursors that poll the connection file
description and call connectionLost of them (and on the connection
itself) after closing the connection from C code.

Fixes #6.