Add support for NOTIFY events.
The connection can register functions that will get called when an
asynchronous NOTIFY is received from the server.

The implementation required a slight change in the way txpostgres uses
the reactor. The Connection object is now always registered as a
reader to receive NOTIFY events as soon as they get sent. This
produces a problem when both the connection and a cursor would be
registered as readers, leading to the connection "stealing" poll
results from the cursor. Specifically, the connection would call
psycopg's poll() and discard the POLL_OK result, which should have
been received by the cursor executing a query and cause it to fire its
Deferred.

To avoid this issue, the connection stops watching the socket while a
cursor is executing a query and resumes watching it once it's done.

Parts of this patch are based on the implementation proposed by Jan
Pobrislo, although the actual code ended up being quite more involved.