Split the poll() method into poll() and continuePolling()
The former is what initiates the process of polling a psycopg2
connection and it returns a Deferred that will get fired when the
poll cycle completes. The latter is responsible for calling down to
psycopg2's C code and depending on the response, either watching the
connection for more IO or firing the Deferred returned from a previous
poll().

The advantage of this is that the polling the connection is sometimes
done without ever goig through user code, like when receiving
asynchronous NOTIFY events. Creating Deferreds there would lead to
error messages being logged if these Deferreds get errbacked, for
instance because the connection to the server got lost.

Another argument for doing it this way is the ability to add a sanity
check in the poll() codepath making sure that the Deferred that's been
returned to the caller is not being clobbered by a second call to
poll() without completing the previous polling cycle.

Patch by Andrey Smirnov with small changes and a unit test by me.