Consume errors from individual connections when starting the pool.
Failures in pool connections were always causing a traceback to be
printed, as observed in #10. Since the individual Deferreds are never
visible to the user, there is nothing that can be done to prevent
that, except for consuming them from withing the start method.

Since gatherResults only got a consumeErrors argument in Twisted
11.1.0 switch to using a DeferredList.

Per report from Marek Kowalski.