Fix errors when connectionFactory is a normal Python method.
Recent versions of psycopg2 (>2.4.2) changed psycopg2.connect to be a
normal Python module method and not a C function exposed from a
module. This made Python treat it as an instance method when it was
assigned as a variable of the Connection class and pass it the object
as the first implicit parameter.

To support both old and new psycopg2 and not break user code that was
inheriting from Connection and setting its own connectionFactory,
force the default one to be a static method.

Bug first reported by Erik Allik, fix proposed by Vasily Evseenko.

Fixes #11.