Make the test for errors in notify observers more robust.
Turns out that the observer that was supposed to raise an exception
was being removed before it had a chance to do its vile work.

Add an extra Deferred to wait upon and add an extra assertion in the
test to make sure that the exception is really raised and logged.