Process NOTIFY events using a Cooperator.
With large numbers of notifies, or with lots of observers, the reactor
thread might be blocked for too long while processing them. Fix this
by using internet.task.cooperate and allowing users to pass their own
Cooperator objects.

While at it add documentation stating that there are not guarantees as
to the order in which observer functions are called and that any
exceptions they might raise are logged and discarded.

Per report from mb5.

Fixes #24.